src/deployable/requirements.txt
argparse
pyyaml
requests